feat!: prevent create command from switching branches (#316)
BREAKING-CHANGE: The create command was previously responsible for
creating and switching branches. To prevent accidentally switching to an
existing branch or accidentally creating a new one when switching was
intended the new `switch` command was added in v1.4. This command should
be used from now on. `create` will only create new branches and fail if
a brach associated with the given ticket already exists (unless
`--reset` is given)